feat: Move and cleanup Connected App in Litify MDAPI
This commit updates the `move_connected_app` task in the `cumulusci.yml` file to use the `MoveAndCleanup` class from the `tasks.move_and_cleanup` module. It moves the Connected App from the `unpackaged/config/litify-mdapi/temp/main/default/connectedApps` directory to the `unpackaged/config/litify-mdapi/src` directory and removes the `unpackaged/config/litify-mdapi/temp` directory.

The purpose of this change is to improve the organization of the Litify MDAPI project by moving the Connected App to the appropriate directory and cleaning up the temporary directory.

Note: This commit message does not include any issue references, tags, or author names.